Edicion de tablas cambiando sus estructuras
Mas precisas donde debe ir enteros, string y otros

+++ models/categoria.go
@@ -1,7 +1,7 @@
 package models
 
 type Categoria struct {
-	ID     int64  `json:"id" gorm:"primary_key;auto_increment"`
+	ID     int16  `json:"id" gorm:"primary_key;auto_increment"`
 	Nombre string `json:"nombre"`
 	Status int    `json:"status"`
 }
