+++ models/producto.go
@@ -1,10 +1,10 @@
 package models
 
 type Producto struct {
-	ID           int64  `json:"id" gorm:"primary_key;auto_increment"`
+	ID           int16  `json:"id" gorm:"primary_key;auto_increment"`
 	Nombre       string `json:"nombre"`
-	Existencia   int    `json:"existencia"`
-	Precio       int    `json:"precio"`
+	Existencia   int16  `json:"existencia"`
+	Precio       int16  `json:"precio"`
 	Status       int    `json:"status"`
 	Id_Categoria uint   `json:"id_categoria"`
 }
